New internal webinos API for intrazone notifications.
The new api 'zonenotification' is hosted by the PZH and then can
be subscribed to by web applications.  This can be used by the
webinos system to trigger notifications on key events, such as
a new user requesting access to the personal zone.

An example function is contained in pzh_webSessionHandling.js
for how to invoke notifications.

Jira issue: WP-523